Change the handling of resources/assets
When the devappserver2 handling of 'pub serve' proxying is removed the
application is in control over all requests. If running in development
mode and 'pub serve' is configured the URL for 'pub serve' is passed in
the environment variablle DART_PUB_SERVE.

This change adds another service available through the context for handling
resources/assets. Through that service the request can serve a static resource.
If running with 'pub serve' this will proxy to 'pub serve' and otherwise serve
from a static file. One just writes:

  context.services.resource.serve();

This is still work in progress, and need at least the following:

  * Access to the resource stream for server side transformation
  * Cofiguring of a default 404 handler
  * Maybe only serve files from buils/web and not web

R=kustermann@google.com
BUG=

Review URL: https://chromereviews.googleplex.com/54507013