Fix async race condition between check of isOpen and making a request
Checking of `connection.isOpen` and invoking `connection.makeRequest`
must be done synchronously.  If there is an `await` in-between, the
connection could have been closed, run out of streams, ...

This CL also introduces a slight delay of 20 ms after connection setup
to allow our client to receive the servers http/2 Settings, which
includes a max-concurrent-streams change of "infinite -> 100".  Without
this delay, we might issue too many RPCs on a connection, just to
receive the limit later and a termination of all streams by the server.