Add 'ClientContext.applicationContext` getter
The application context class holds useful information,
such as the app id, version, etc. Currently, there's no
way to get at this information, as it's not exposed in
the `package:appengine`  API.  This PR exposes it via
the `ClientContext` API.

The use case is apps that run different versions or
project instances need a way to get that info. In the
case of the Flutter dashboard backend, it needs  to
construct a serialized proto containing the app id.