Instantiate new DatastoreDB for each request.
Before this change, servers that used DatastroeDB were succeptible to
concurrency issues, where different concurrentrequests would share the
same DatastoreDB instance. Specifically, the issue was as follows:

appengine registers a single gcloud:DatastoreDB instance upon server startup
 ↳ gcloud:DatastoreDB maintains a reference to a gcloud:Datastore
   (implemented by appengine:GrpcDatastoreImpl)
 ↳ appengine:GrpcDatastoreImpl maintains a reference to grpc:ClientChannel
 ↳ grpc:ClientChannel maintains a reference to an grpc:Http2ClientConnection
 ↳ grpc:Http2ClientConnection maintains state

This would cause issues whereby the client connection would throw for request
B because it would think that it was in a closing statedue to a call from
request A.

To fix this, this change ensires that each request gets a new instance
of DatastoreDB.

This is a breaking change because it changes the interface of the
`ContextRegistry` constructor.